+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 
 0.2.5 (YYYY-MM-DD)
 ------------------
+* Remove deadlock in TableProxy weakref.finalize on Python 3.6 (:pr:`113`)
 * Use python-casacore wheels for travis testing, instead of kernsuite packages (:pr:`115`)
 
 0.2.4 (2020-04-24)